+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added more detailed examples to the forecast module documentation.
 - Improved handling of datatime indexes in DataFrames on forecast module.
 - Added warning messages if passed list values contains non numeric items.
+- Added missing unittests for forecast module with request.get dependencies using MagicMock.
+- Added the Solar.Forecast method.
 
 ## [0.3.19] - 2022-09-14
 ### Fix